Create aboutIPP.html
Created new page for about us containing history of IPP as well as the
Policies of the company.
May extract these policies into their own pages later on, but would like
to condense the content for now.
Updated the styles to improve margins for text to ensure no overlay from
social media buttons.
Included about us in the navigation bar.