Created troubleshooting.html
Used an accordian style panel for the troubleshooting questions rather
than laying all information out on the page for a cleaner approach.
Added the installation videos from youtube to the bottom of the page
with responsive sizing, but not too happy with how they're coming out
due to the low resolution so may come back and resolve this.
Two tickets raised for some descriptions of the troubleshooting which
requires the owners to follow up - incomplete sentences, descriptions
missing.